ScamGuard_AI_Bot/requirements.txt
# ScamGuard AI Bot - Dependencies
# Install with: pip install -r requirements.txt

# Telegram Bot API
python-telegram-bot[ext]==21.*

# AI / LLM Integration
openai>=1.0.0
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Ollama (runs locally, no pip install needed)
# Download from: https://ollama.ai

# Optional: WhatsApp Business API
# requests>=2.31.0  # Already available via httpx/openai deps